Changed imports to absolute paths; renamed utils.py to query_db.py
